Create SVM_LampActivation.py
This code uses the SVM model developed with Matlab's Classification Learner tools to identify if the patient is in either "relaxation" or "pain" state and activate the lamp's color (green or red, respectively). Beforehand, you need to run the "ArduinoIDE_DataAdq" code for Arduino IDE; you can find it in this repository's main.